added new class with methods of the game, rewrited a start new game method
